requirement.txt
streamlit
requests
pandas
ipykernel
plotly_express
geopy